selector folder index & selectorOptions to file
